refactor event data into its own module
